requirements.txt
langchain-openai==0.2.10
langchain-core==0.3.28
langchain==0.3.7
openai==1.57.2
python-dotenv==1.0.1
pytest==8.3.4
google-generativeai==0.8.3
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic-settings==2.1.0
